+++ tests/adapter/mongo/test_variant_events.py
@@ -16,8 +16,8 @@ def test_mark_causative(adapter, institute_obj, case_obj, user_obj, variant_obj)
     adapter.user_collection.insert_one(user_obj)
     adapter.variant_collection.insert_one(variant_obj)
 
-    assert adapter.variant_collection.find().count() > 0
-    assert adapter.event_collection.find().count() == 0
+    assert sum(1 for i in adapter.variant_collection.find()) > 0
+    assert sum(1 for i in adapter.event_collection.find()) == 0
 
     variant = adapter.variant_collection.find_one()
 
@@ -33,7 +33,7 @@ def test_mark_causative(adapter, institute_obj, case_obj, user_obj, variant_obj)
     # THEN the case should have a causative variant
     assert len(updated_case['causatives']) == 1
     # THEN two events should have been created, one for the case and one for the variant
-    assert adapter.event_collection.find().count() == 2
+    assert sum(1 for i in adapter.event_collection.find()) == 2
 
     # THEN assert that case status is updated to solved
     assert updated_case['status'] == 'solved'
@@ -77,7 +77,7 @@ def test_unmark_causative(adapter, institute_obj, case_obj, user_obj, variant_ob
     assert updated_case['status'] == 'active'
     ## THEN assert that two more events was created
 
-    assert adapter.event_collection.find().count() == 4
+    assert sum(1 for i in adapter.event_collection.find()) == 4
 
 
 def test_order_verification(adapter, institute_obj, case_obj, user_obj, variant_obj):
@@ -88,8 +88,8 @@ def test_order_verification(adapter, institute_obj, case_obj, user_obj, variant_
     adapter.user_collection.insert_one(user_obj)
     adapter.variant_collection.insert_one(variant_obj)
 
-    assert adapter.variant_collection.find().count() > 0
-    assert adapter.event_collection.find().count() == 0
+    assert sum(1 for i in adapter.variant_collection.find()) > 0
+    assert sum(1 for i in adapter.event_collection.find()) == 0
 
     variant = adapter.variant_collection.find_one()
     assert variant.get('sanger_ordered') is not True
@@ -105,7 +105,7 @@ def test_order_verification(adapter, institute_obj, case_obj, user_obj, variant_
     )
 
     # THEN one events should have been created, one for the variant
-    assert adapter.event_collection.find().count() == 2
+    assert sum(1 for i in adapter.event_collection.find()) == 2
 
     # THEN updated variant should have same id as original variant
     assert variant.get('_id') == updated_variant.get('_id')
@@ -127,8 +127,8 @@ def test_cancel_verification(adapter, institute_obj, case_obj, user_obj, variant
     adapter.user_collection.insert_one(user_obj)
     adapter.variant_collection.insert_one(variant_obj)
 
-    assert adapter.variant_collection.find().count() > 0
-    assert adapter.event_collection.find().count() == 0
+    assert sum(1 for i in adapter.variant_collection.find()) > 0
+    assert sum(1 for i in adapter.event_collection.find()) == 0
 
     link = 'orderSangerlink'
     updated_variant = adapter.order_verification(
@@ -168,8 +168,8 @@ def test_dismiss_variant(adapter, institute_obj, case_obj, user_obj, variant_obj
     adapter.user_collection.insert_one(user_obj)
     adapter.variant_collection.insert_one(variant_obj)
 
-    assert adapter.variant_collection.find().count() > 0
-    assert adapter.event_collection.find().count() == 0
+    assert sum(1 for i in adapter.variant_collection.find()) > 0
+    assert sum(1 for i in adapter.event_collection.find()) == 0
 
     variant = adapter.variant_collection.find_one()