+++ tests/commands/view/test_view_hpo_cmd.py
@@ -13,10 +13,10 @@ def test_view_hpo(mock_app):
     result =  runner.invoke(cli, ['view', 'hpo'])
     assert result.exit_code == 0
     # no term shold be returned
-    assert "No matching terms found" in result.output
+    assert "Found 0 terms" in result.output
 
     # no HPO terms in database, so insert one
-    assert store.hpo_term_collection.find().count() == 0
+    assert store.hpo_term_collection.find_one() is None
     hpo_term = {
         '_id' : 'HP:0000311',
         'hpo_id' : 'HP:0000311',